supporting proper find in anchored widgets

+++ src/ct/ct_types.h
@@ -467,7 +467,7 @@ struct CtAnchMatch {
     int             start_offset;
     Glib::ustring   line_content;
     CtAnchWidgType  anch_type;
-    int             anch_cell_idx;
+    size_t          anch_cell_idx;
     int             anch_offs_start;
     int             anch_offs_end;
 };
